fix: replace %zu with %u under newlib nano formatting
CONFIG_NEWLIB_NANO_FORMAT=y is set project-wide (sdkconfig.defaults), so
the nano printf can't handle %zu/%zd/%zx — it prints the specifier
literally and arg-shifts the rest (this is what hid the real camera frame
size as "len=zu, format=<len value>"). Sweep the remaining occurrences,
casting size_t args to unsigned with %u.

Shared paths (affect every board):
- audio/demuxer/ogg_demuxer.cc (x3)
- boards/common/afsk_demod.cc
- boards/common/esp_video.cc (P4 camera upload)

Board-local:
- boards/sensecap-watcher/sscma_camera.cc
- boards/otto-robot/websocket_control_server.cc (x3)

Verified: xiao-esp32-s3-sense builds clean; waveshare/esp32-s3-cam (the
other official camera board, shares the common Esp32Camera path) passes
the compile-only board smoke test.

+++ main/boards/otto-robot/websocket_control_server.cc
@@ -124,7 +124,7 @@ void WebSocketControlServer::HandleMessage(httpd_req_t *req, const char* data, s
     }
     
     if (len > 4096) {
-        ESP_LOGE(TAG, "Message too long: %zu bytes", len);
+        ESP_LOGE(TAG, "Message too long: %u bytes", (unsigned)len);
         return;
     }
     
@@ -177,14 +177,14 @@ void WebSocketControlServer::AddClient(httpd_req_t *req) {
     int sock_fd = httpd_req_to_sockfd(req);
     if (clients_.find(sock_fd) == clients_.end()) {
         clients_[sock_fd] = req;
-        ESP_LOGI(TAG, "Client connected: %d (total: %zu)", sock_fd, clients_.size());
+        ESP_LOGI(TAG, "Client connected: %d (total: %u)", sock_fd, (unsigned)clients_.size());
     }
 }
 
 void WebSocketControlServer::RemoveClient(httpd_req_t *req) {
     int sock_fd = httpd_req_to_sockfd(req);
     clients_.erase(sock_fd);
-    ESP_LOGI(TAG, "Client disconnected: %d (total: %zu)", sock_fd, clients_.size());
+    ESP_LOGI(TAG, "Client disconnected: %d (total: %u)", sock_fd, (unsigned)clients_.size());
 }
 
 size_t WebSocketControlServer::GetClientCount() const {
